Refactored to be decoupled
Refactored the overall project to be decoupled, to allow for higher levels of flexibility and correctness.
