+++ docs/jwhitt3r/test_repo/doc1.md
@@ -1,0 +1,4 @@
+This is another documentation file
+[TestLink]("https://thisisatestlink.co.uk")
+[I'm an inline-style link with title](https://www.google.com "Google's Homepage")
+[testBroken404](https://github.com/jwhitt3r3)
